reduced memory requirements
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@16944 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: 4f9a6e505102c7815a6e3728de289d210dceda6d

+++ world_models/topological_map/include/topological_map/region.h
@@ -43,9 +43,9 @@ using std::ostream;
 
 struct Cell2D
 {
-  int r, c;
+  short r, c;
   Cell2D() {}
-  Cell2D(const int row, const int column) : r(row), c(column) {}
+  Cell2D(const short row, const short column) : r(row), c(column) {}
 };
 
 ostream& operator<< (ostream& str, const Cell2D& c);
